+++ src/aws/variables.tf
@@ -154,3 +154,9 @@ variable "scw_group_id" {
   type = string
   default = ""
 }
+
+variable "deletion_retention_period" {
+  description = "The time duration (in days) an object will exist in the 'deleted' directory"
+  type = number
+  default = 21
+}