+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add scaleway storage configuration
 - Add Peertube pipeline to VOD
 - Celery task queue
+- Refactor video S3 deletion
 
 ### Changed